Updates to base, basic and all imports
